requirements.txt
bigquery-python==0.1.1
